REFACTOR :: 폴더 구조
