+++ apps/user/src/components/main/FaqItem/FaqItem.tsx
