+++ apps/user/src/components/main/NoticeItem/NoticeItem.tsx
