+++ apps/user/src/components/main/Schedule/Schedule.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import ScheduleItem from './ScheduleItem';
+import ScheduleItem from '../ScheduleItem/ScheduleItem';
 import { color, font } from '@maru/theme';
 import { flex } from '@maru/utils';
